requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.5.0.5
celery==4.2.0
cffi==1.15.0
Django==3.2.12
django-celery-beat==1.1.1
djangorestframework==3.13.1
dnspython==2.2.0
eventlet==0.33.0
greenlet==1.1.2
importlib-metadata==4.11.1
kombu==4.6.11
pycparser==2.21
pytz==2021.3
rabbitmq==0.2.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
vine==1.3.0
zipp==3.7.0
